KOffice Libs -complation fixes for msvc -one fix for gcc 3.3.x.
svn path=/trunk/koffice/; revision=684826

+++ libs/kwmf/kowmfreadprivate.cc
@@ -1183,8 +1183,8 @@ void KoWmfReadPrivate::xyToAngle( int xStart, int yStart, int xEnd, int yEnd, in
 {
     double aStart, aLength;
 
-    aStart = atan2( yStart,  xStart );
-    aLength = atan2( yEnd, xEnd ) - aStart;
+    aStart = atan2( (double)yStart,  (double)xStart );
+    aLength = atan2( (double)yEnd, (double)xEnd ) - aStart;
 
     angleStart = (int)((aStart * 2880) / 3.14166);
     angleLength = (int)((aLength * 2880) / 3.14166);
